+++ Pages/Models/Post.cs
@@ -3,6 +3,7 @@ namespace hopefullyAWebForum.Pages.Models;
 public class Post
 {
     public int PostId { get; set; }
+    public int UserId { get; set; }
     public string AuthorName { get; set; } // UserId -> join
     public string Topic { get; set; }
     public string Content { get; set; }