+++ Pages/Models/User.cs
@@ -3,6 +3,7 @@ namespace hopefullyAWebForum.Pages.Models;
 public class User
 {
     public int Id { get; set; }
-    public string Username {get; set;}
-    public string Password {get; set;}
+    public string Username { get; set; }
+    public DateTime JoinDate { get; set; }
 }
+